update timeseries data class
